+++ src/PhpSpreadsheet/Calculation/Calculation.php
@@ -224,6 +224,16 @@ class Calculation
             'functionCall' => 'acosh',
             'argumentCount' => '1',
         ],
+        'ACOT' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'ACOT'],
+            'argumentCount' => '1',
+        ],
+        'ACOTH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'ACOTH'],
+            'argumentCount' => '1',
+        ],
         'ADDRESS' => [
             'category' => Category::CATEGORY_LOOKUP_AND_REFERENCE,
             'functionCall' => [LookupRef::class, 'cellAddress'],
@@ -460,6 +470,16 @@ class Calculation
             'functionCall' => 'cosh',
             'argumentCount' => '1',
         ],
+        'COT' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'COT'],
+            'argumentCount' => '1',
+        ],
+        'COTH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'COTH'],
+            'argumentCount' => '1',
+        ],
         'COUNT' => [
             'category' => Category::CATEGORY_STATISTICAL,
             'functionCall' => [Statistical::class, 'COUNT'],
@@ -524,6 +544,16 @@ class Calculation
             'category' => Category::CATEGORY_STATISTICAL,
             'functionCall' => [Statistical::class, 'CRITBINOM'],
             'argumentCount' => '3',
+        ],
+        'CSC' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'CSC'],
+            'argumentCount' => '1',
+        ],
+        'CSCH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'CSCH'],
+            'argumentCount' => '1',
         ],
         'CUBEKPIMEMBER' => [
             'category' => Category::CATEGORY_CUBE,
@@ -1663,6 +1693,16 @@ class Calculation
             'category' => Category::CATEGORY_TEXT_AND_DATA,
             'functionCall' => [TextData::class, 'SEARCHINSENSITIVE'],
             'argumentCount' => '2,3',
+        ],
+        'SEC' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'SEC'],
+            'argumentCount' => '1',
+        ],
+        'SECH' => [
+            'category' => Category::CATEGORY_MATH_AND_TRIG,
+            'functionCall' => [MathTrig::class, 'SECH'],
+            'argumentCount' => '1',
         ],
         'SECOND' => [
             'category' => Category::CATEGORY_DATE_AND_TIME,